+++ src/minter/src/sol_rpc_client/requests.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct GetSignaturesForAddressRequestOptions {
-    pub limit: Option<u64>,
+    pub limit: Option<u8>,
     pub commitment: Option<String>,
     pub until: Option<String>,
     pub before: Option<String>,
